:sparkles: use async and channels to handle and propagate events (#24)
* :recycle: change panic hook and log
* :recycle: use channels to deliver events

+++ src/components/networks.rs
@@ -11,6 +11,7 @@ use ratatui::style::{Style, Stylize};
 use ratatui::text::{Line, Span};
 use ratatui::widgets::{Block, Borders, Clear, Padding, Paragraph, TableState, Wrap};
 use ratatui::Frame;
+use tokio::sync::mpsc::UnboundedSender;
 
 use crate::action::Action;
 use crate::components::Component;
@@ -33,6 +34,7 @@ pub struct Networks {
     state: TableState,
     networks: Vec<[String; 4]>,
     show_popup: Popup,
+    action_tx: Option<UnboundedSender<Action>>,
 }
 
 impl Networks {
@@ -42,6 +44,7 @@ impl Networks {
             state: Default::default(),
             networks: Vec::new(),
             show_popup: Popup::None,
+            action_tx: None,
         }
     }
 
@@ -119,9 +122,13 @@ impl Component for Networks {
         "Networks"
     }
 
-    fn update(&mut self, action: Option<Action>) -> Result<Option<Action>> {
+    fn register_action_handler(&mut self, action_tx: UnboundedSender<Action>) {
+        self.action_tx = Some(action_tx);
+    }
+
+    fn update(&mut self, action: Action) -> Result<()> {
         match action {
-            Some(Action::Refresh) => {
+            Action::Tick => {
                 let options = ListNetworksOptions {
                     filters: self.filters.clone(),
                 };
@@ -145,41 +152,33 @@ impl Component for Networks {
                         })
                         .collect()
                 });
-                Ok(None)
             }
-            Some(Action::Down) => {
+            Action::Down => {
                 self.next();
-                Ok(None)
             }
-            Some(Action::Up) => {
+            Action::Up => {
                 self.previous();
-                Ok(None)
             }
-            Some(Action::Delete) => {
+            Action::Delete => {
                 if let Some(id) = self.get_selected_network_info() {
                     self.show_popup = Popup::Delete(id);
-                    Ok(Some(Action::Refresh))
-                } else {
-                    Ok(None)
                 }
             }
-            Some(Action::Ok) => {
-                let show_popup = &self.show_popup;
-                match show_popup {
-                    Popup::Delete(id) => {
-                        delete_network(id)?;
-                        self.show_popup = Popup::None;
-                        self.update(Some(Action::Refresh))
+            Action::Ok => {
+                if let Popup::Delete(id) = &self.show_popup {
+                    delete_network(id)?;
+                    self.show_popup = Popup::None;
+                    if let Some(tx) = self.action_tx.clone() {
+                        tx.send(Action::Tick)?;
                     }
-                    _ => Ok(None),
                 }
             }
-            Some(Action::PreviousScreen) => {
+            Action::PreviousScreen => {
                 self.show_popup = Popup::None;
-                Ok(Some(Action::Refresh))
             }
-            _ => Ok(action),
-        }
+            _ => {}
+        };
+        Ok(())
     }
 
     fn draw(&mut self, f: &mut Frame<'_>, area: Rect) {
@@ -187,7 +186,7 @@ impl Component for Networks {
             .constraints([Constraint::Percentage(100)])
             .split(area);
         let t = table(
-            self.get_name(),
+            self.get_name().to_string(),
             ["Id", "Name", "Size", "Age"],
             self.networks.clone(),
             &NETWORK_CONSTRAINTS,
